cli/requirements.txt
boto3==1.12.47
prompt_toolkit==2.0.7
sty==1.0.0b8
click==7.1.2
pyreadline==2.1
pyhcl==0.4.4
python-consul==1.1.0
tqdm==4.45.0
npyscreen==4.10.5
pexpect==4.8.0
mfa==0.1.1
beautifulsoup4==4.9.0
keyring==21.2.0
keyrings.alt==3.4.0
pytz==2020.1
websocket_client==0.57.0
zipstream==1.1.4
tabulate==0.8.7
